+++ src/renderer/src/page/latest-anime/index.tsx
@@ -1,9 +1,0 @@
-import { RouterLayout } from '@renderer/components/layout/root/RouterLayout'
-
-export default function LastAnime() {
-  return (
-    <RouterLayout>
-      <div className="flex h-full items-center justify-center">还在编写中，敬请期待</div>
-    </RouterLayout>
-  )
-}