+++ src/renderer/src/router/name.ts
@@ -1,5 +1,4 @@
 export enum RouteName {
   PLAYER = '/player',
-  LATEST_ANIME = '/latest-anime',
   LIBRARY = '/library',
 }